pyproject.toml
[project]
name = "gork"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.3",
    "valkey-glide>=2.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]
